+++ src/packages/v2-plugin/parsers.ts
@@ -1,19 +1,50 @@
-import type { Parser } from 'prettier';
+import { parseLineByLineAndAssemble } from 'core-parts';
+import type { Parser, ParserOptions } from 'prettier';
+import { format } from 'prettier';
 import { parsers as babelParsers } from 'prettier/parser-babel';
 import { parsers as htmlParsers } from 'prettier/parser-html';
 import { parsers as typescriptParsers } from 'prettier/parser-typescript';
 
+const addon = {
+  parseBabel: (text: string, options: ParserOptions) =>
+    babelParsers.babel.parse(text, { babel: babelParsers.babel }, options),
+  parseTypescript: (text: string, options: ParserOptions) =>
+    typescriptParsers.typescript.parse(text, { typescript: typescriptParsers.typescript }, options),
+};
+
+function transformParser(
+  parserName: 'babel' | 'typescript' | 'vue',
+  defaultParser: Parser,
+): Parser {
+  return {
+    ...defaultParser,
+    parse: (text: string, parsers: { [parserName: string]: Parser }, options: ParserOptions) => {
+      const formattedText = format(text, {
+        ...options,
+        plugins: [],
+        endOfLine: 'lf',
+      });
+      const ast = defaultParser.parse(formattedText, { [parserName]: defaultParser }, options);
+
+      const result = parseLineByLineAndAssemble(
+        formattedText,
+        ast,
+        // @ts-ignore
+        options,
+        addon,
+      );
+
+      return {
+        type: 'FormattedText',
+        body: result,
+      };
+    },
+    astFormat: 'brace-style-ast',
+  };
+}
+
 export const parsers: { [parserName: string]: Parser } = {
-  babel: {
-    ...babelParsers.babel,
-    astFormat: 'babel-ast',
-  },
-  typescript: {
-    ...typescriptParsers.typescript,
-    astFormat: 'typescript-ast',
-  },
-  vue: {
-    ...htmlParsers.vue,
-    astFormat: 'vue-ast',
-  },
+  babel: transformParser('babel', babelParsers.babel),
+  typescript: transformParser('typescript', typescriptParsers.typescript),
+  vue: transformParser('vue', htmlParsers.vue),
 };
